+++ leap_ec/binary_rep/problems.py
@@ -12,18 +12,11 @@ from leap_ec.problem import ScalarProblem
 ##############################
 class MaxOnes(ScalarProblem):
     """
-    Implementation of MAX ONES problem where the individuals are represented
-    by a bit vector
-
-    We don't need an encoder since the raw genome is *already* in the
-    phenotypic space.
+    Implementation of the classic max-ones problem, where the individuals
+    are represented by a bit vector
     """
 
     def __init__(self, maximize=True):
-        """
-        Create a MAX ONES problem with individuals that have bit vectors of
-        size `length`
-        """
         super().__init__(maximize)
 
     def evaluate(self, individual):
@@ -40,6 +33,53 @@ class MaxOnes(ScalarProblem):
             raise ValueError(("Expected phenome to be a numpy array. "
                               f"Got {type(individual.phenome)}."))
         return np.count_nonzero(individual.phenome == 1)
+
+
+##############################
+# Class DeceptiveTrap
+##############################
+class DeceptiveTrap(ScalarProblem):
+    """
+    A simple bi-modal function whose global optimum is the Boolean vector
+    of all 1's, but in which fitness *decreases* as the number of 1's in
+    the vector *increases*—giving it a local optimum of [0, ..., 0] with a 
+    very wide basin of attraction.
+
+    :param int dimensions: number of dimensions we expect to find in the bitstring.
+    """
+    def __init__(self, maximize=True):
+        super().__init__(maximize=maximize)
+    
+    def evaluate(self, individual):
+        
+        """
+        >>> from leap_ec.individual import Individual
+        >>> import numpy as np
+        >>> p = DeceptiveTrap()
+
+        The trap function has a global maximum when the number of one's
+        is maximized:
+
+        >>> ind = Individual(np.array([1, 1, 1, 1, 1, 1, 1, 1, 1, 1]))
+        >>> p.evaluate(ind)
+        10
+
+        It's minimized when we have just one zero:
+        >>> ind = Individual(np.array([1, 1, 1, 1, 0, 1, 1, 1, 1, 1]))
+        >>> p.evaluate(ind)
+        0
+
+        And has a local optimum when we have no ones at all:
+        >>> ind = Individual(np.array([0, 0, 0, 0, 0, 0, 0, 0, 0, 0]))
+        >>> p.evaluate(ind)
+        9
+        """
+        dimensions = len(individual.phenome)
+        max_ones = np.count_nonzero(individual.phenome == 1)
+        if max_ones == dimensions:
+            return dimensions
+        else:
+            return dimensions - max_ones - 1
 
 
 ##############################